src/requirements.txt
# Python application required libraries
tornado
paho-mqtt
